+++ Aero.Infrastructure/Persistences/Entities/Strike.cs
@@ -6,10 +6,10 @@ namespace Aero.Infrastructure.Persistences.Entities
 {
     public sealed class Strike : BaseEntity,IDeviceId
     {
-        public int module_id { get; set; }
+        public short module_id { get; set; }
         public Module module { get; set; }
         public Door strike_door { get; set; }
-        public int door_id { get; set; }
+        public short door_id { get; set; }
         public short output_no { get; set; }
         public short relay_mode { get; set; }
         public short offline_mode { get; set; }
@@ -19,7 +19,7 @@ namespace Aero.Infrastructure.Persistences.Entities
         public int device_id { get; set; }
         public Device device { get; set; }
 
-        public Strike(int device,int moduleid,int doorid,short outputno,short relaymode,short offlinemode,short strikmax,short strikemin,short strikemode,int location) : base(location) 
+        public Strike(int device,short moduleid,short doorid,short outputno,short relaymode,short offlinemode,short strikmax,short strikemin,short strikemode,int location) : base(location) 
         {
             this.device_id = device;
             this.module_id = moduleid;
